# heuristic: pre-select records by package-name