Might have gotten carried away and spent some time this morning looking at the performance: worst offender is `_test_patch_yaml` with ~75% time spent there. However didn't find an obvious way to optimize the function, just seems to get called a lot.

This simple heuristic brings down the number of calls to `_test_patch_yaml` by pre-selecting index records based on exact package name matches. On my M2 this saves almost a factor of two if you are willing to compromise a bit on basic math 😛

**before:** python recipe/show_diff.py  1302.53s user 194.61s system 276% cpu 9:00.76 total
**after:** python recipe/show_diff.py  747.86s user 121.13s system 249% cpu 5:48.66 total

It's a heuristic, but it's simple enough that it might be worth it.

<details>
<summary>colorful icy flamegraph </summary>
<img width="1078" alt="image" src="https://github.com/conda-forge/conda-forge-repodata-patches-feedstock/assets/15384900/5d444116-e01f-49a4-959e-6f1f52648aa5">
</details>

<details>
<summary>_test_patch_yaml profiling</summary>

```
   Total time: 299.785 s
File: /Users/christian/repo/conda-forge-repodata-patches-feedstock/recipe/patch_yaml_utils.py
Function: _test_patch_yaml at line 142

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   142                                           @profile
   143                                           def _test_patch_yaml(patch_yaml, record, subdir, fn):
   144 111649440    7226035.0      0.1      2.4      keep = True
   145 118836270   26080006.0      0.2      8.7      for k, v in patch_yaml["if"].items():
   146 117339564   21357372.0      0.2      7.1          if k.startswith("not_"):
   147    379885      56631.0      0.1      0.0              k = k[4:]
   148    379885      37366.0      0.1      0.0              neg = True
   149                                                   else:
   150 116959679    9587079.0      0.1      3.2              neg = False
   151                                           
   152 117339564   24030407.0      0.2      8.0          if k in record:
   153  92281437   11905612.0      0.1      4.0              if k == "version" and not any(
   154      3696        333.0      0.1      0.0                  symb in v for symb in ["*", "[", "]", "?", "(", ")"]
   155                                                       ):
   156      3327      57227.0     17.2      0.0                  _keep = parse_version(record[k]) == parse_version(v)
   157                                                       else:
   158  92274414   41853820.0      0.5     14.0                  _keep = fnmatch(str(record[k]), str(v))
   159                                           
   160  25061823    2682156.0      0.1      0.9          elif k == "subdir_in":
   161   3227997    2683200.0      0.8      0.9              _keep = _fnmatch_str_or_list(subdir, v)
   162                                           
   163  21833826    4994282.0      0.2      1.7          elif k[-3:] in ["_lt", "_le", "_gt", "_ge", "_eq", "_ne"] and (
   164   5141798    1382242.0      0.3      0.5              k[:-3] in record
   165     44896       7075.0      0.2      0.0              or k[:-3] in ["timestamp"]  # some records do not have a timestamp
   166                                                   ):
   167   5141798     908642.0      0.2      0.3              subk = k[:-3]
   168                                           
   169                                                       # some records do not have some keys, so we can
   170                                                       # special case that here
   171   5141798     556223.0      0.1      0.2              if subk == "timestamp":
   172   5135690    1039330.0      0.2      0.3                  rv = record.get(subk, 0)
   173                                                       else:
   174      6108        963.0      0.2      0.0                  rv = record[subk]
   175                                           
   176   5141798     616891.0      0.1      0.2              if subk == "version":
   177      6016      42995.0      7.1      0.0                  rv = parse_version(rv)
   178      6016      38771.0      6.4      0.0                  v = parse_version(v)
   179   5135782     569487.0      0.1      0.2              elif subk in ["build_number", "timestamp"]:
   180   5135782     636119.0      0.1      0.2                  rv = int(rv)
   181   5135782     572086.0      0.1      0.2                  v = int(v)
   182                                           
   183   5141798     853127.0      0.2      0.3              op = k[-2:]
   184   5141798     579332.0      0.1      0.2              if op == "lt":
   185   5136479     606261.0      0.1      0.2                  _keep = rv < v
   186      5319        530.0      0.1      0.0              elif op == "le":
   187      2050        650.0      0.3      0.0                  _keep = rv <= v
   188      3269        315.0      0.1      0.0              elif op == "gt":
   189        28         12.0      0.4      0.0                  _keep = rv > v
   190      3241        297.0      0.1      0.0              elif op == "ge":
   191      3141       1266.0      0.4      0.0                  _keep = rv >= v
   192       100         25.0      0.2      0.0              elif op == "eq":
   193       100         64.0      0.6      0.0                  _keep = rv == v
   194                                                       elif op == "ne":
   195                                                           _keep = rv != v
   196                                           
   197  16692028    3731763.0      0.2      1.2          elif k.endswith("_in") and k[:-3] in record:
   198   3003343     402341.0      0.1      0.1              subk = k[:-3]
   199   3003343    4521881.0      1.5      1.5              _keep = _fnmatch_str_or_list(record[subk], v)
   200                                           
   201  13688685    4121313.0      0.3      1.4          elif k.startswith("has_") and k[len("has_") :] in ["depends", "constrains"]:
   202  12929165    2110299.0      0.2      0.7              subk = k[len("has_") :]
   203  12929165    1586254.0      0.1      0.5              if not isinstance(v, list):
   204  10817622    1047160.0      0.1      0.3                  v = [v]
   205  25858330   30097495.0      1.2     10.0              _keep = all(
   206  12929165    1230885.0      0.1      0.4                  any(fnmatch(dep, _v) for dep in record.get(subk, [])) for _v in v
   207                                                       )
   208                                           
   209    759520      79437.0      0.1      0.0          elif k == "artifact_in":
   210    759520    1165330.0      1.5      0.4              _keep = _fnmatch_str_or_list(fn, v)
   211                                           
   212                                                   else:
   213                                                       raise KeyError("Unrecognized 'if' key '%s'!" % k)
   214                                           
   215 117339564   38061677.0      0.3     12.7          if neg:
   216    379885      47799.0      0.1      0.0              keep = keep and (not _keep)
   217                                                   else:
   218 116959679   12305375.0      0.1      4.1              keep = keep and _keep
   219                                           
   220 117339564   13251915.0      0.1      4.4          if not keep:
   221 110152734    9701013.0      0.1      3.2              break
   222                                           
   223 111649440   15358415.0      0.1      5.1      return keep

```
</details>